hotel_management/requirements.txt
django-environ
djangorestframework==3.16.0 
djangorestframework-simplejwt
django-redis
django-cors-headers
drf-yasg[swagger-ui]
python-dotenv==1.0.0
django==4.2
mysqlclient==2.2.0
python-jose[cryptography]==3.3.0

django-summernote

requests
djangorestframework-camel-case

unidecode
user-agents
ua-parser
pyyaml
Pillow
redis
shortuuid
google-auth
celery
holidays
confluent_kafka

channels 
uvicorn
channels-redis